backend/requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy[asyncio]==2.0.25
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.5
psycopg2-binary==2.9.9

# Supabase
supabase>=2.10.0

# OpenAI
openai>=1.30.0

# Utilities
python-dotenv==1.0.0
python-multipart==0.0.6
httpx>=0.25.0

# CORS
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Production
gunicorn==21.2.0

tiktoken>=0.5.0